requirements.txt
python-dotenv==1.0.0
rich==13.7.0
